requirements.txt
opencv-python
deepface
tf-keras
streamlit
plotly
numpy
tensorflow-macos
tensorflow-metal
